refactored tests into correct package
