+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/stores/FederatedStoreTypeTest.java
@@ -15,12 +15,10 @@
  */
 
 
-package uk.gov.gchq.gaffer.gaas.factories;
+package uk.gov.gchq.gaffer.gaas.stores;
 
 import org.junit.jupiter.api.Test;
 import uk.gov.gchq.gaffer.common.model.v1.GafferSpec;
-import uk.gov.gchq.gaffer.gaas.stores.AbstractStoreTypeBuilder;
-import uk.gov.gchq.gaffer.gaas.stores.FederatedStoreType;
 import uk.gov.gchq.gaffer.gaas.utilities.UnitTest;
 import static org.junit.jupiter.api.Assertions.assertEquals;
